# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WeedTown API",
    "version": "2.4.0",
    "description": "API de WeedTown, red social de la comunidad cannábica de México. Identidad federada con Mastodon (OAuth 2.0) y sesión mediante JWT propio en el header Authorization. Seguridad: headers via helmet, CORS restringido al frontend, rate limiting (300 peticiones/15 min general; 20/15 min en el flujo OAuth), payload JSON limitado a 100 kB e imágenes a 5 MB. El perfil público (GET /api/profile/{id}) no expone PII (email, teléfono, nombre real, edad, nacimiento ni género). Las rutas de mercado y admin son stubs pendientes de implementación."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por el backend tras el flujo OAuth de Mastodon (expira en 7 días)."
      }
    },
...
  },
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Moderación"
        ],
        "summary": "Buscar cuentas para moderar",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Sin `q` devuelve las cuentas suspendidas ahora mismo.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 2
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          }
        }
      }
    },
...
    "/api/reports": {
      "post": {
        "tags": [
          "Reportes"
        ],
        "summary": "Reportar contenido o una cuenta",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Idempotente: reportar dos veces lo mismo no duplica la cola, solo actualiza el motivo. El chat privado NO es reportable — moderarlo obligaría a leer mensajes 1 a 1. Límite de 20 reportes por hora.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetType",
                  "targetId",
                  "reason"
                ],
                "properties": {
                  "targetType": {
                    "type": "string",
                    "enum": [
                      "POST",
                      "COMMENT",
                      "FORUM_POST",
                      "FORUM_COMMENT",
                      "USER",
                      "SUBFORUM"
                    ]
                  },
                  "targetId": {
                    "type": "integer"
                  },
                  "reason": {
                    "type": "string",
                    "enum": [
                      "SPAM",
                      "ACOSO",
                      "ODIO",
                      "ILEGAL",
                      "DESINFORMACION",
                      "SEXUAL",
                      "SUPLANTACION",
                      "OTRO"
                    ]
                  },
                  "detail": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reporte registrado (`duplicado: true` si ya lo habías reportado)"
          },
          "400": {
            "description": "Tipo no reportable, motivo inválido, o es tu propio contenido"
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe lo que intentas reportar"
          },
          "429": {
            "description": "Demasiados reportes en poco tiempo"
          }
        }
      }
    },
...
    "/api/admin/reports": {
      "get": {
        "tags": [
          "Moderación"
        ],
        "summary": "Cola de revisión",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "El contenido reportado viaja con el reporte para no tener que abrir otra pantalla, junto con el historial de la cuenta señalada (cuántos reportes acumula). Nunca incluye quién reportó.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "ACCIONADO",
                "DESCARTADO"
              ],
              "default": "PENDIENTE"
            }
          },
          {
            "name": "reason",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SPAM",
                "ACOSO",
                "ODIO",
                "ILEGAL",
                "DESINFORMACION",
                "SEXUAL",
                "SUPLANTACION",
                "OTRO"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          }
        }
      }
    },
    "/api/admin/reports/{id}/descartar": {
      "post": {
        "tags": [
          "Moderación"
        ],
        "summary": "El reporte no procede",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Sale de la cola sin tocar el contenido ni la cuenta, pero queda en el historial para dar contexto si vuelve a aparecer.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          },
          "409": {
            "description": "Ese reporte ya estaba resuelto"
          }
        }
      }
    },
    "/api/admin/content/{type}/{id}/ocultar": {
      "post": {
        "tags": [
          "Moderación"
        ],
        "summary": "Ocultar contenido (reversible)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Deja de verse para toda la comunidad, incluido su autor, pero la fila se conserva con quién y cuándo lo ocultó. Resuelve solo los reportes pendientes sobre ese objeto y notifica a su autor con el motivo — nunca con quién reportó ni qué moderador actuó.",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "POST",
                "COMMENT",
                "FORUM_POST",
                "FORUM_COMMENT"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "enum": [
                      "SPAM",
                      "ACOSO",
                      "ODIO",
                      "ILEGAL",
                      "DESINFORMACION",
                      "SEXUAL",
                      "SUPLANTACION",
                      "OTRO"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          },
          "409": {
            "description": "Ese contenido ya estaba oculto"
          }
        }
      }
    },
    "/api/admin/content/{type}/{id}/mostrar": {
      "post": {
        "tags": [
          "Moderación"
        ],
        "summary": "Restaurar contenido oculto",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "POST",
                "COMMENT",
                "FORUM_POST",
                "FORUM_COMMENT"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          },
          "409": {
            "description": "Ese contenido no estaba oculto"
          }
        }
      }
    },
    "/api/admin/users/{id}/suspender": {
      "post": {
        "tags": [
          "Moderación"
        ],
        "summary": "Suspender una cuenta temporalmente",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Bloquea publicar, comentar, chatear, mandar toques y subir imágenes. NO bloquea leer: es una pausa, no una expulsión. Caduca sola en la fecha indicada. Un MOD solo puede suspender cuentas con rol USER; suspender a un MOD o ADMIN requiere ser ADMIN.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "days",
                  "reason"
                ],
                "properties": {
                  "days": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365
                  },
                  "reason": {
                    "type": "string",
                    "enum": [
                      "SPAM",
                      "ACOSO",
                      "ODIO",
                      "ILEGAL",
                      "DESINFORMACION",
                      "SEXUAL",
                      "SUPLANTACION",
                      "OTRO"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Duración o motivo inválidos, o intento de suspenderte a ti"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          }
        }
      }
    },
    "/api/admin/users/{id}/levantar": {
      "post": {
        "tags": [
          "Moderación"
        ],
        "summary": "Levantar una suspensión antes de tiempo",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          }
        }
      }
    },
...
    "/api/admin/stats": {
      "get": {
        "tags": [
          "Moderación"
        ],
        "summary": "Panorama del panel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Tu cuenta no tiene rol de moderación"
          }
        }
      }
    },
...
  }
}